+++ test/RemembermeTest.php
@@ -40,17 +40,20 @@ class RemembermeTest extends PHPUnit_Framework_TestCase
 
   /* Basic cases */
 
-  public function testReturnFalseIfNoCookieExists()
+  public function testNoCookieExists()
   {
-    $this->assertFalse($this->rememberme->login());
+      $this->assertFalse($this->rememberme->login()->isSuccess());
+      $this->assertFalse($this->rememberme->login()->cookieExists());
   }
 
   public function testReturnFalseIfCookieIsInvalid()
   {
-    $this->cookie->method("getValue")->willReturn("DUMMY");
-    $this->assertFalse($this->rememberme->login());
+      $this->cookie->method("getValue")->willReturn("DUMMY");
+      $this->assertFalse($this->rememberme->login()->isSuccess());
+      $this->assertTrue($this->rememberme->login()->hasPossibleManipulation());
       $this->cookie->method("getValue")->willReturn($this->userid."|a");
-    $this->assertFalse($this->rememberme->login());
+      $this->assertFalse($this->rememberme->login()->isSuccess());
+      $this->assertTrue($this->rememberme->login()->hasPossibleManipulation());
   }
 
   public function testLoginTriesToFindTripletWithValuesFromCookie() {
@@ -67,7 +70,7 @@ class RemembermeTest extends PHPUnit_Framework_TestCase
 
   /* Success cases */
 
-  public function testReturnTrueIfTripletIsFound() {
+  public function testSuccessIsTrueIfTripletIsFound() {
       $this->cookie->method("getValue")->willReturn( implode("|", array(
           $this->userid,
           $this->validToken,
@@ -77,8 +80,21 @@ class RemembermeTest extends PHPUnit_Framework_TestCase
     $this->storage->expects($this->once())
       ->method("findTriplet")
       ->will($this->returnValue(Birke\Rememberme\Storage\StorageInterface::TRIPLET_FOUND));
-    $this->assertEquals($this->userid, $this->rememberme->login());
+    $this->assertTrue($this->rememberme->login()->isSuccess());
   }
+
+  public function testCredentialsAreInResultIfTripletIsFound() {
+        $this->cookie->method("getValue")->willReturn( implode("|", array(
+            $this->userid,
+            $this->validToken,
+            $this->validPersistentToken
+        )));
+
+        $this->storage->expects($this->once())
+            ->method("findTriplet")
+            ->will($this->returnValue(Birke\Rememberme\Storage\StorageInterface::TRIPLET_FOUND));
+        $this->assertEquals($this->userid, $this->rememberme->login()->getCredential());
+   }
 
   public function testStoreNewTripletInCookieIfTripletIsFound() {
     $oldcookieValue = implode("|", array(
@@ -159,36 +175,41 @@ class RemembermeTest extends PHPUnit_Framework_TestCase
 
   /* Failure Cases */
 
-  public function testFalseIfTripletIsNotFound() {
+  public function testResultIndicatesExpiredWhenTripletIsNotFound() {
       $this->cookie->method("getValue")->willReturn( implode("|", array(
       $this->userid, $this->validToken, $this->validPersistentToken)));
+      $this->storage->expects($this->once())
+          ->method("findTriplet")
+          ->will($this->returnValue(Birke\Rememberme\Storage\StorageInterface::TRIPLET_NOT_FOUND));
 
-    $this->storage->expects($this->once())
-      ->method("findTriplet")
-      ->will($this->returnValue(Birke\Rememberme\Storage\StorageInterface::TRIPLET_NOT_FOUND));
-    $this->assertFalse($this->rememberme->login());
+      $result = $this->rememberme->login();
+
+      $this->assertFalse($result->isSuccess());
+      $this->assertTrue($result->isExpired());
   }
 
-  public function testFalseIfTripletIsInvalid() {
+  public function testResultIndicatesManipulationIfTripletIsInvalid() {
       $this->cookie->method("getValue")->willReturn( implode("|", array(
       $this->userid, $this->invalidToken, $this->validPersistentToken)));
+      $this->storage->expects($this->once())
+          ->method("findTriplet")
+          ->will($this->returnValue(Birke\Rememberme\Storage\StorageInterface::TRIPLET_INVALID));
 
-    $this->storage->expects($this->once())
-      ->method("findTriplet")
-      ->will($this->returnValue(Birke\Rememberme\Storage\StorageInterface::TRIPLET_INVALID));
-    $this->assertFalse($this->rememberme->login());
+      $result = $this->rememberme->login();
+
+      $this->assertFalse($result->isSuccess());
+      $this->assertTrue($result->hasPossibleManipulation());
   }
 
   public function testCookieIsExpiredIfTripletIsInvalid() {
       $this->cookie->method("getValue")->willReturn( implode("|", array(
       $this->userid, $this->invalidToken, $this->validPersistentToken)));
-    $now = time();
-    $this->storage->expects($this->once())
-      ->method("findTriplet")
-      ->will($this->returnValue(Birke\Rememberme\Storage\StorageInterface::TRIPLET_INVALID));
-    $this->cookie->expects($this->once())
-      ->method("deleteCookie");
-    $this->rememberme->login();
+      $this->storage->expects($this->once())
+          ->method("findTriplet")
+          ->will($this->returnValue(Birke\Rememberme\Storage\StorageInterface::TRIPLET_INVALID));
+      $this->cookie->expects($this->once())
+          ->method("deleteCookie");
+      $this->rememberme->login();
   }
 
   public function testAllStoredTokensAreClearedIfTripletIsInvalid() {
@@ -204,18 +225,6 @@ class RemembermeTest extends PHPUnit_Framework_TestCase
     $this->rememberme->login();
     $this->rememberme->setCleanStoredTokensOnInvalidResult(false);
     $this->rememberme->login();
-  }
-
-  public function testInvalidTripletStateIsStored() {
-      $this->cookie->method("getValue")->willReturn( implode("|", array(
-      $this->userid, $this->invalidToken, $this->validPersistentToken)));
-
-    $this->storage->expects($this->once())
-      ->method("findTriplet")
-      ->will($this->returnValue(Birke\Rememberme\Storage\StorageInterface::TRIPLET_INVALID));
-    $this->assertFalse($this->rememberme->loginTokenWasInvalid());
-    $this->rememberme->login();
-    $this->assertTrue($this->rememberme->loginTokenWasInvalid());
   }
 
   /* Salting test */
@@ -237,17 +246,6 @@ class RemembermeTest extends PHPUnit_Framework_TestCase
     );
     $this->rememberme->setSalt($salt);
     $this->rememberme->login();
-  }
-
-  public function testSaltIsAddedToTokensOnCookieIsValid() {
-    $salt = "Mozilla Firefox 4.0";
-      $this->cookie->method("getValue")->willReturn( implode("|", array(
-      $this->userid, $this->validToken, $this->validPersistentToken)));
-    $this->storage->expects($this->once())
-      ->method("findTriplet")
-      ->with($this->equalTo($this->userid), $this->equalTo($this->validToken.$salt), $this->equalTo($this->validPersistentToken.$salt));
-    $this->rememberme->setSalt($salt);
-    $this->rememberme->cookieIsValid($this->userid);
   }
 
   public function testSaltIsAddedToTokensOnCreateCookie() {